lib: repositories setting code refactor
In order to avoid specifying repeteadly "--repository" I tried to optimize it by using arrays. This, of course, implies changing the interpreter of the script to bash, but that's not a big issue considering both mklive and mkiso already are like this.

Also, this avoids declaring repositories that don't apply for the target architechture.

Lastly, the repositories declaring was moved into a function, so they can be set in a specific stage and not necessarily at the beginning when calling lib.sh.

With this, I'll be removing the repository declaring from mklive.sh and hence, mkiso.sh, since I consider redundant specifiying the same repositories trice.