mkiso: code refactoring
In a similar way as mklive, now the package list is an array. This makes the process much easier, readable and less redundant (I hope). Some variables like ARCH_PKGS, REPO_NONFREE and so on were dropped thanks to this.

THEMES_PKGS was dropped in favor of cereus-themes-presets.

The path handling for the locales for the module of the same name for Calamares was moved to mklive and now it replaces against @@LOCALES_FILE@@ on the target dir.

Arrays like CEREUS_BASEPKGS and XORG_PKGS are specified globally while excluding base variant.

Repository declaring was dropped in favor of the approach explained in the previous commit, which was continued here for both mkiso and mklive.

shellcheck warnings were addresed.

I know this commit includes much more changes that it should, but there were a lot of things to fix and wasn't sure about how to split it. I have to improve that aspect. Of course, there are a lot more of changes pending, like the package list for every variant.